kas: migrate Google Repo to Siemens KAS
The use of Google Repo introduced multiple issues related to version
management, complex CI, and overall dev experience. Siemens KAS provides
a more reliable and maintainable solution for our needs.

Note that during the migration it was detected that meta-ros is not used
in the build process, therefore it was removed. Moreover, the
meta-synaptics branch has been switched to a static SHA.