bsp: recipes-kernel: linux: sbc: fix board boot after reset
The AstraSBC may not boot properly after reset (power toggle needed) if
M2 BT related UART2 is enabled. The root cause of this issue is not
known yet but it is triggered if pinmux of SM_SPI2_SS3n line is
configured to UART_RXD.

Therefore disable UART2 and BT functionality for now to make board boot
reliable.